Move engine helpers into `engine/helpers.ts`
